test: Move instruments' tests in a single scope
