+++ tests/instruments/test_bluefors.py
