+++ tests/instruments/test_erasynth.py
