+++ tests/instruments/test_oscillator.py
